LU-3852 tests: skip tests with large file when no room
Strengthen sanity-hsm sub-tests using large files by skipping
then if not enough space instead to fail.

Also re-enable test_251, and test_[200,221,223b] (for LU-4178),
who have all been wrongly disabled because of their multiple
failures on an auto-test node with very low disk space for the
Servers VMs!!

Also fix a bug in sanity/test_56w that should not be run if
$OSTCOUNT < 2, that was triggered due to my very restictive
test environment osscount/ostcount=1 to limit the Lustre FS size
as much as possible.

Test-Parameters: ostcount=1 ostsizegb=1 envdefinitions=SLOW=yes
Test-Parameters: osscount=1 mdscount=1 testlist=sanity-hsm
Reviewed-on: http://review.whamcloud.com/12456
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/tests/sanity-hsm.sh
@@ -13,10 +13,6 @@ export PATH=$PWD/$SRCDIR:$SRCDIR:$PWD/$SRCDIR/utils:$PATH:/sbin:/usr/sbin
 ONLY=${ONLY:-"$*"}
 # bug number for skipped test:    3815     3939
 ALWAYS_EXCEPT="$SANITY_HSM_EXCEPT 34 35 36 40"
-# bug number for skipped test:4178
-ALWAYS_EXCEPT="$ALWAYS_EXCEPT 200 221 223b"
-# bug number for skipped test:LU-3852
-ALWAYS_EXCEPT="$ALWAYS_EXCEPT 251"
 # UPDATE THE COMMENT ABOVE WITH BUG NUMBERS WHEN CHANGING ALWAYS_EXCEPT!
 
 LUSTRE=${LUSTRE:-$(cd $(dirname $0)/..; echo $PWD)}
@@ -589,11 +585,23 @@ cleanup_large_files() {
 	[ $ratio -gt 50 ] && find $MOUNT -size +10M -exec rm -f {} \;
 }
 
+check_enough_free_space() {
+	local nb=$1
+	local unit=$2
+	local need=$((nb * unit /1024))
+	local free=$(df -kP $MOUNT | tail -1 | awk '{print $4}')
+	(( $need >= $free )) && return 1
+	return 0
+}
+
 make_large_for_striping() {
 	local file2=${1/$DIR/$DIR2}
 	local sz=$($LCTL get_param -n lov.*-clilov-*.stripesize | head -n1)
 
 	cleanup_large_files
+
+	check_enough_free_space 5 $sz
+	[ $? != 0 ] && return $?
 
 	dd if=/dev/urandom of=$file2 count=5 bs=$sz conv=fsync ||
 		error "cannot create $file2"
@@ -604,6 +612,9 @@ make_large_for_progress() {
 	local file2=${1/$DIR/$DIR2}
 
 	cleanup_large_files
+
+	check_enough_free_space 39 1000000
+	[ $? != 0 ] && return $?
 
 	# big file is large enough, so copy time is > 30s
 	# so copytool make 1 progress
@@ -619,6 +630,9 @@ make_large_for_progress_aligned() {
 
 	cleanup_large_files
 
+	check_enough_free_space 33 1048576
+	[ $? != 0 ] && return $?
+
 	# big file is large enough, so copy time is > 30s
 	# so copytool make 1 progress
 	# size is a multiple of 1M to have stripe
@@ -632,6 +646,9 @@ make_large_for_cancel() {
 	local file2=${1/$DIR/$DIR2}
 
 	cleanup_large_files
+
+	check_enough_free_space 103 1048576
+	[ $? != 0 ] && return $?
 
 	# Copy timeout is 100s. 105MB => 105s
 	dd if=/dev/urandom of=$file2 count=103 bs=1M conv=fsync ||
@@ -1116,6 +1133,8 @@ test_12c() {
 	local f=$DIR/$tdir/$tfile
 	$LFS setstripe -c 2 $f
 	local fid=$(make_large_for_striping $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	local FILE_CRC=$(md5sum $f)
 
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
@@ -1991,6 +2010,8 @@ test_26() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
 	wait_request_state $fid ARCHIVE SUCCEED
 
@@ -2028,6 +2049,8 @@ test_27b() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
 	wait_request_state $fid ARCHIVE SUCCEED
 	$LFS hsm_release $f
@@ -2047,6 +2070,8 @@ test_28() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
 	wait_request_state $fid ARCHIVE SUCCEED
 
@@ -2225,6 +2250,8 @@ test_31b() {
 
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
 	wait_request_state $fid ARCHIVE SUCCEED
 	$LFS hsm_release $f
@@ -2246,6 +2273,8 @@ test_31c() {
 
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress_aligned $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
 	wait_request_state $fid ARCHIVE SUCCEED
 	$LFS hsm_release $f
@@ -2267,6 +2296,8 @@ test_33() {
 
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
 	wait_request_state $fid ARCHIVE SUCCEED
 	$LFS hsm_release $f
@@ -2331,6 +2362,8 @@ test_34() {
 
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
 	wait_request_state $fid ARCHIVE SUCCEED
 	$LFS hsm_release $f
@@ -2364,6 +2397,8 @@ test_35() {
 	local f=$DIR/$tdir/$tfile
 	local f1=$DIR/$tdir/$tfile-1
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	local fid1=$(copy_file /etc/passwd $f1)
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
 	wait_request_state $fid ARCHIVE SUCCEED
@@ -2400,6 +2435,8 @@ test_36() {
 
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
 	wait_request_state $fid ARCHIVE SUCCEED
 	$LFS hsm_release $f
@@ -2593,6 +2630,7 @@ test_56() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
 
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f ||
 		error "could not archive file"
@@ -2715,6 +2753,7 @@ test_60() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
 
 	local mdtidx=0
 	local mdt=${MDT_PREFIX}${mdtidx}
@@ -2837,6 +2876,7 @@ test_71() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
 
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f ||
 		error "could not archive file"
@@ -3113,6 +3153,8 @@ test_104() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	# if cdt is on, it can serve too quickly the request
 	cdt_disable
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER --data $DATA $f
@@ -3424,6 +3466,8 @@ test_200() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_cancel $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	# test with cdt on is made in test_221
 	cdt_disable
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
@@ -3465,6 +3509,8 @@ test_202() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
+
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
 	wait_request_state $fid ARCHIVE SUCCEED
 
@@ -3510,6 +3556,7 @@ test_221() {
 
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_cancel $f)
+	[ $? != 0 ] && skip "not enough free space" && return
 
 	changelog_setup
 
@@ -3617,6 +3664,7 @@ test_223b() {
 
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
 
 	changelog_setup
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
@@ -3677,6 +3725,7 @@ test_225() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_progress $f)
+	[ $? != 0 ] && skip "not enough free space" && return
 
 	changelog_setup
 	$LFS hsm_archive --archive $HSM_ARCHIVE_NUMBER $f
@@ -3880,6 +3929,7 @@ test_251() {
 	mkdir -p $DIR/$tdir
 	local f=$DIR/$tdir/$tfile
 	local fid=$(make_large_for_cancel $f)
+	[ $? != 0 ] && skip "not enough free space" && return
 
 	cdt_disable
 	# to have a short test
